requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
Pillow==10.4.0
imagehash==4.3.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
httpx==0.27.0
python-dotenv==1.0.1
exifread==3.0.0
aiofiles==24.1.0
beautifulsoup4==4.12.3
lxml==5.3.0
